☩マップ文字もでたぞ☩
