+++ Plugin/mapView.cpp
@@ -1,221 +1,347 @@
 #include "stdinc.h"
 #include "byte_pattern.h"
 
-namespace MapView {
+
+namespace MapView
+{
 	/*-----------------------------------------------*/
 
-	errno_t stackSizeChange_hook(CK2Version version) {
-		std::string desc = "stack size change";
-
-		switch (version) {
-		case v2_8_X:
-		case v3_0_X:
-			// sub esp,178h
-			byte_pattern::temp_instance().find_pattern("81 EC 78 01 00 00 8B 43 18");
-			if (byte_pattern::temp_instance().has_size(1,desc)) {
-				injector::WriteMemory<uint8_t>(
-					byte_pattern::temp_instance().get_first().address(2),
-					0x7A,
-					true
-				);
-			}
-			else return CK2ERROR1;
-			return NOERROR;
-		}
-
-		return CK2ERROR1;
-	}
-
-	/*-----------------------------------------------*/
-
-	uintptr_t m_5;
-	__declspec(naked) void m_1()
+	uintptr_t p_5;
+	__declspec(naked) void p_1()
 	{
 		__asm {
-			cmp byte ptr[eax + edx], ESCAPE_SEQ_1;
-			jz m_10;
+			cmp byte ptr[eax + ebx], ESCAPE_SEQ_1;
+			jz p_10;
 
-			cmp byte ptr[eax + edx], ESCAPE_SEQ_2;
-			jz m_11;
+			cmp byte ptr[eax + ebx], ESCAPE_SEQ_2;
+			jz p_11;
 
-			cmp byte ptr[eax + edx], ESCAPE_SEQ_3;
-			jz m_12;
+			cmp byte ptr[eax + ebx], ESCAPE_SEQ_3;
+			jz p_12;
 
-			cmp byte ptr[eax + edx], ESCAPE_SEQ_4;
-			jz m_13;
+			cmp byte ptr[eax + ebx], ESCAPE_SEQ_4;
+			jz p_13;
 
-			jmp m_3;
+			mov al, [eax + ebx];
+			lea ecx, [ebp - 0xFC];
+			mov byte ptr[ebp - 0x34], al;
 
-		m_10:
-			movzx eax, word ptr[eax + edx + 1];
-			jmp m_1x;
+			push p_5;
+			ret;
 
-		m_11:
-			movzx eax, word ptr[eax + edx + 1];
+		p_10:
+			movzx eax, word ptr[eax + ebx + 1];
+			lea ecx, [ebp - 0x34];
+			mov byte ptr[ecx], ESCAPE_SEQ_1;
+			mov[ecx + 1], ax;
+			jmp p_1x;
+
+		p_11:
+			movzx eax, word ptr[eax + ebx + 1];
+			lea ecx, [ebp - 0x34];
+			mov byte ptr[ecx], ESCAPE_SEQ_2;
+			mov[ecx + 1], ax;
+			jmp p_1x;
+
+		p_12:
+			movzx eax, word ptr[eax + ebx + 1];
+			lea ecx, [ebp - 0x34];
+			mov byte ptr[ecx], ESCAPE_SEQ_3;
+			mov[ecx + 1], ax;
+			jmp p_1x;
+
+		p_13:
+			movzx eax, word ptr[eax + ebx + 1];
+			lea ecx, [ebp - 0x34];
+			mov byte ptr[ecx], ESCAPE_SEQ_4;
+			mov[ecx + 1], ax;
+
+		p_1x:
+			lea ecx, [ebp - 0xFC];
+
+			push p_5;
+			ret;
+		}
+	}
+
+	uintptr_t q_4;
+	__declspec(naked) void q_1()
+	{
+		__asm {
+			cmp byte ptr[eax], ESCAPE_SEQ_1;
+			jz q_2;
+
+			cmp byte ptr[eax], ESCAPE_SEQ_2;
+			jz q_2;
+
+			cmp byte ptr[eax], ESCAPE_SEQ_3;
+			jz q_2;
+
+			cmp byte ptr[eax], ESCAPE_SEQ_4;
+			jz q_2;
+
+			jmp q_3;
+
+		q_2:
+			mov ecx, [ebp - 0x34];
+			mov[eax], ecx;
+			mov byte ptr[eax + 0x10], 3;
+
+		q_3:
+			push 0xFFFFFFFF;
+			push 0;
+			push eax;
+
+			push q_4;
+			ret;
+		}
+	}
+
+	/*-----------------------------------------------*/
+
+	errno_t func1_2_hook(CK2Version version) {
+		std::string desc = "func1,2";
+
+		switch (version) {
+		case v2_8_X:
+		case v3_0_X:
+			// mov al,[eax+ebx]
+			byte_pattern::temp_instance().find_pattern("8A 04 18 8D 8D 04 FF");
+			if (byte_pattern::temp_instance().has_size(1, desc)) {
+				injector::MakeJMP(byte_pattern::temp_instance().get_first().address(), p_1);
+				p_5 = byte_pattern::temp_instance().get_first().address(12);
+	
+				injector::MakeJMP(byte_pattern::temp_instance().get_first().address(0x14), q_1);
+				q_4 = byte_pattern::temp_instance().get_first().address(0x19);
+			}
+			else return CK2ERROR1;
+			return NOERROR;
+		}
+		return CK2ERROR1;
+	}
+
+	/*-----------------------------------------------*/
+
+	uintptr_t o_5;
+	__declspec(naked) void o_1()
+	{
+		__asm {
+			cmp byte ptr[eax + ebx], ESCAPE_SEQ_1;
+			jz o_10;
+
+			cmp byte ptr[eax + ebx], ESCAPE_SEQ_2;
+			jz o_11;
+
+			cmp byte ptr[eax + ebx], ESCAPE_SEQ_3;
+			jz o_12;
+
+			cmp byte ptr[eax + ebx], ESCAPE_SEQ_4;
+			jz o_13;
+
+			jmp o_3;
+
+		o_10:
+			movzx eax, word ptr[eax + ebx + 1];
+			jmp o_1x;
+
+		o_11:
+			movzx eax, word ptr[eax + ebx + 1];
 			sub eax, SHIFT_2;
-			jmp m_1x;
+			jmp o_1x;
 
-		m_12:
-			movzx eax, word ptr[eax + edx + 1];
+		o_12:
+			movzx eax, word ptr[eax + ebx + 1];
 			add eax, SHIFT_3;
-			jmp m_1x;
+			jmp o_1x;
 
-		m_13:
-			movzx eax, word ptr[eax + edx + 1];
+		o_13:
+			movzx eax, word ptr[eax + ebx + 1];
 			add eax, SHIFT_4;
-
-		m_1x:
-			add edx, 2;
-			mov[ebp - 0x24], edx;
-			sub edx, 2;
-
+	
+		o_1x:
 			movzx eax, ax;
 			cmp eax, NO_FONT;
-			ja m_4;
+			ja o_2;
 			mov eax, NOT_DEF;
-			jmp m_4;
 
-		m_3:
-			movzx eax, byte ptr[eax + edx];
+		o_2:
+			add ebx, 2;
+			jmp o_4;
 
-		m_4:
-			mov esi, [esi + eax * 4];
+		o_3:
+			movzx eax, byte ptr[eax + ebx];
 
-			mov[ebp - 0x172], ax;
-			cmp word ptr[ebp - 0x172], 0x100;
+		o_4:
+			mov edx, [ebp - 0x20];
 
-			jb m_421;
-			cmp word ptr[ebp - 0x1C], 3;
-
-			ja m_421;
-			inc edx;
-			mov word ptr[ebp - 0x172], 0x10;
-
-		m_421:
-			push m_5;
+			push o_5;
 			ret;
 		}
 	}
 
 	/*-----------------------------------------------*/
 
-	errno_t view_hook(CK2Version version) {
-		std::string desc = "map view";
+	errno_t func3_hook(CK2Version version) {
+		std::string desc = "func3";
 
 		switch (version) {
 		case v2_8_X:
 		case v3_0_X:
-			// movzx eax,byte ptr [eax+edx]
-			byte_pattern::temp_instance().find_pattern("0F B6 04 10 8B 34 86");
-			if (byte_pattern::temp_instance().has_size(1, desc)) {
-				injector::MakeJMP(byte_pattern::temp_instance().get_first().address(), m_1);
-				m_5 = byte_pattern::temp_instance().get_first().address(7);
-			}
-			else return CK2ERROR1;
-			return NOERROR;
-		}
-
-		return CK2ERROR1;
-	}
-
-	/*-----------------------------------------------*/
-
-	uintptr_t kk_4;
-	__declspec(naked) void kk_1() {
-		__asm {
-			cmp word ptr[ebp - 0x172], 0x100;
-			jb kk_2;
-			sub eax, 3;
-			jmp kk_3;
-
-		kk_2:
-			dec eax;
-
-		kk_3:
-			sqrtss xmm1, xmm1;
-
-			push kk_4;
-			ret;
-		}
-
-	};
-
-	/*-----------------------------------------------*/
-
-	errno_t mapAdj1_hook(CK2Version version) {
-		std::string desc = "map adj 1";
-
-		switch (version) {
-		case v2_8_X:
-		case v3_0_X:
-			//dec eax
-			byte_pattern::temp_instance().find_pattern("48 F3 0F 51 C9 66 0F 6E");
-			if (byte_pattern::temp_instance().has_size(1, desc)) {
-				injector::MakeJMP(byte_pattern::temp_instance().get_first().address(), kk_1);
-				kk_4 = byte_pattern::temp_instance().get_first().address(5);
-			}
-			else return CK2ERROR1;
-			return NOERROR;
-		}
-
-		return CK2ERROR1;
-	}
-
-	/*-----------------------------------------------*/
-
-	uintptr_t jj_2;
-	__declspec(naked) void jj_1() {
-		__asm {
-
-			mov eax, [ebp - 0x0B4];
-			mov esi, [ebx + 0x38];
-			mov edx, [ebp - 0x24];
-			inc edx;
-
-			push jj_2;
-			ret;
-		}
-	};
-
-	/*-----------------------------------------------*/
-
-	errno_t mapAdj2_hook(CK2Version version) {
-		std::string desc = "map adj 2";
-
-		switch (version) {
-		case v2_8_X:
-		case v3_0_X:
-			// mov eax, [ebp-0B4h]
-			byte_pattern::temp_instance().find_pattern("8B 85 4C FF FF FF 8B 73");
-			if (byte_pattern::temp_instance().has_size(1, desc)) {
-				injector::WriteMemory<uint8_t>(byte_pattern::temp_instance().get_first().address(-0xE), 0x90, true);
-				injector::MakeJMP(byte_pattern::temp_instance().get_first().address(), jj_1);
-				jj_2 = byte_pattern::temp_instance().get_first().address(9);
-			}
-			else return CK2ERROR1;
-			return NOERROR;
-		}
-
-		return CK2ERROR1;
-	}
-
-	/*-----------------------------------------------*/
-
-	errno_t stackClear_hook(CK2Version version) {
-		std::string desc = "stack clear";
-
-		switch (version){
-		case v2_8_X:
-		case v3_0_X:
-			// lea esp,[ebp-180h]
-			byte_pattern::temp_instance().find_pattern("8D A5 80 FE FF FF 5F");
+			// movzx eax, byte ptr [eax+ebx]
+			byte_pattern::temp_instance().find_pattern("0F B6 04 18 8B 55 E0");
 			if (byte_pattern::temp_instance().has_size(1,desc)) {
-				injector::WriteMemory<uint8_t>(byte_pattern::temp_instance().get_first().address(2), 0x7E, true);
+				injector::MakeJMP(byte_pattern::temp_instance().get_first().address(), o_1);
+				o_5 = byte_pattern::temp_instance().get_first().address(7);
 			}
 			else return CK2ERROR1;
 			return NOERROR;
 		}
+		return CK2ERROR1;
+	}
 
+	/*-----------------------------------------------*/
+
+	uintptr_t n_4_v28;
+	__declspec(naked) void n_1_v28()
+	{
+		__asm {
+			cmp byte ptr[eax + ecx], ESCAPE_SEQ_1;
+			jz n_10;
+
+			cmp byte ptr[eax + ecx], ESCAPE_SEQ_2;
+			jz n_11;
+
+			cmp byte ptr[eax + ecx], ESCAPE_SEQ_3;
+			jz n_12;
+
+			cmp byte ptr[eax + ecx], ESCAPE_SEQ_4;
+			jz n_13;
+
+			jmp n_3;
+
+		n_10:
+			movzx eax, word ptr[eax + ecx + 1];
+			jmp n_1x;
+
+		n_11:
+			movzx eax, word ptr[eax + ecx + 1];
+			sub eax, SHIFT_2;
+			jmp n_1x;
+
+		n_12:
+			movzx eax, word ptr[eax + ecx + 1];
+			add eax, SHIFT_3;
+			jmp n_1x;
+
+		n_13:
+			movzx eax, word ptr[eax + ecx + 1];
+			add eax, SHIFT_4;
+
+		n_1x:
+			add ecx, 2;
+			movzx eax, ax;
+			mov[ebp - 0x14], ecx;
+
+			cmp eax, NO_FONT;
+			ja n_2;
+			mov eax, NOT_DEF;
+
+		n_2:
+			jmp n_5;
+
+		n_3:
+			movzx eax, byte ptr[eax + ecx];
+		n_5:
+			mov edx, [edx + eax * 4 + 0x8C];
+
+			push n_4_v28;
+			ret;
+		}
+	}
+
+	uintptr_t n_4_v30;
+	__declspec(naked) void n_1_v30()
+	{
+		__asm {
+			cmp byte ptr[eax + ecx], ESCAPE_SEQ_1;
+			jz n_10;
+
+			cmp byte ptr[eax + ecx], ESCAPE_SEQ_2;
+			jz n_11;
+
+			cmp byte ptr[eax + ecx], ESCAPE_SEQ_3;
+			jz n_12;
+
+			cmp byte ptr[eax + ecx], ESCAPE_SEQ_4;
+			jz n_13;
+
+			jmp n_3;
+
+		n_10:
+			movzx eax, word ptr[eax + ecx + 1];
+			jmp n_1x;
+
+		n_11:
+			movzx eax, word ptr[eax + ecx + 1];
+			sub eax, SHIFT_2;
+			jmp n_1x;
+
+		n_12:
+			movzx eax, word ptr[eax + ecx + 1];
+			add eax, SHIFT_3;
+			jmp n_1x;
+
+		n_13:
+			movzx eax, word ptr[eax + ecx + 1];
+			add eax, SHIFT_4;
+
+		n_1x:
+			add ecx, 2;
+			movzx eax, ax;
+			mov[ebp - 0x14], ecx;
+
+			cmp eax, NO_FONT;
+			ja n_2;
+			mov eax, NOT_DEF;
+
+		n_2:
+			jmp n_5;
+
+		n_3:
+			movzx eax, byte ptr[eax + ecx];
+		n_5:
+			mov edx, [edx + eax * 4 + 0xA4]; // 0x8C -> 0xA4
+
+			push n_4_v30;
+			ret;
+		}
+	}
+
+
+	/*-----------------------------------------------*/
+
+	errno_t func4_hook(CK2Version version) {
+		std::string desc = "func4";
+
+		switch (version) {
+		case v2_8_X:
+			byte_pattern::temp_instance().find_pattern("0F B6 04 08 8B 94 82 8C");
+			if (byte_pattern::temp_instance().has_size(1, desc)) {
+				injector::MakeJMP(byte_pattern::temp_instance().get_first().address(), n_1_v28);
+				n_4_v28 = byte_pattern::temp_instance().get_first().address(11);
+			}
+			else return CK2ERROR1;
+			return NOERROR;
+
+		case v3_0_X:
+			byte_pattern::temp_instance().find_pattern("0F B6 04 08 8B 94 82 A4");
+			if (byte_pattern::temp_instance().has_size(1, desc)) {
+				injector::MakeJMP(byte_pattern::temp_instance().get_first().address(), n_1_v30);
+				n_4_v30 = byte_pattern::temp_instance().get_first().address(11);
+			}
+			else return CK2ERROR1;
+			return NOERROR;
+		}
 		return CK2ERROR1;
 	}
 
@@ -224,19 +350,18 @@ namespace MapView {
 	errno_t init(CK2Version version) {
 		errno_t result = NOERROR;
 
-		byte_pattern::debug_output2("map font view");
+		byte_pattern::debug_output2("map adj");
 
-		// M用にスタック修正
-		result |= stackSizeChange_hook(version);
-		// マップフォント表示処理
-		result |= view_hook(version);
-		//マップテキストの表示位置調整
-		result |= mapAdj1_hook(version);
-		//マップフォントの位置調整
-		result |= mapAdj2_hook(version);
-		// スタック後始末
-		result |= stackClear_hook(version);
+		/* */
+		result |= func1_2_hook(version);
+
+		/* */
+		result |= func3_hook(version);
+
+		/* */
+		result |= func4_hook(version);
 
 		return result;
 	}
+
 }
